rag_system/requirements.txt
# Core RAG Components
transformers>=4.35.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4  # Use faiss-gpu for GPU acceleration
torch>=2.0.0

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
python-pptx>=0.6.23
markdown>=3.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Text Processing
nltk>=3.8.0
spacy>=3.7.0
tiktoken>=0.5.0

# Retrieval
rank-bm25>=0.2.2

# API & Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# UI
streamlit>=1.28.0
plotly>=5.18.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
numpy>=1.24.0
pandas>=2.1.0

# Logging & Monitoring
loguru>=0.7.0

# Database (optional)
sqlalchemy>=2.0.0
redis>=5.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
